docker/seed_cell_size_2d/requirements.txt
chardet==2.2.1
click==6.7
cycler==0.10.0
dask==0.9.0
decorator==4.0.9
dtool==0.12.1
iniparse==0.4
jicbioimage.core==0.15.0
jicbioimage.illustrate==0.6.1
jicbioimage.segment==0.4.0
jicbioimage.transform==0.6.0
Jinja2==2.9.6
kitchen==1.1.1
lxml==3.2.1
MarkupSafe==1.0
matplotlib==1.5.1
networkx==1.11
numpy==1.11.0
Pillow==3.2.0
pycurl==7.19.0
pygpgme==0.3
pyliblzma==0.5.3
pyparsing==2.1.3
python-dateutil==2.5.3
python-magic==0.4.13
pytz==2016.4
pyxattr==0.5.1
PyYAML==3.12
scikit-image==0.12.3
scipy==0.17.0
six==1.10.0
toolz==0.7.4
urlgrabber==3.10
yum-metadata-parser==1.1.4
